корзина доработки

+++ client/src/servises/redux/slice/cartSlice.ts
@@ -3,21 +3,41 @@ import { IReqCart } from "@types";
 import CartEntities from "domain/entities/CartEntities/Cart.entities";
 import { RequestCart } from "servises/repository/Axios/Request";
 import { RTKCart } from "servises/repository/RTK/RTKCart";
+import { AppDispatch, RootState } from '../createStore'
 
 const cartAdapter = createEntityAdapter<IReqCart>({
   selectId: (product) => product.id,
 })
 
-export const cartSelector = cartAdapter.getSelectors((state: any) => state.cart)
+export const cartSelector = cartAdapter.getSelectors((state: RootState) => state.cart)
 
 
+
+export const fetchAllCart = createAsyncThunk(
+  'cart/getAll',
+  async (_, {dispatch,getState,rejectWithValue }) => {
+    try {
+      const state = getState()
+      
+      const request = await RequestCart.allCart()
+      if (request.status == 200 && request.data) {
+        dispatch(addAllCart(request.data.cart))
+        dispatch(setTotalPrice(request.data.totalPrice))
+      }
+      
+    } catch (error:any) {
+      return rejectWithValue(error.response.data)
+    }
+  }
+)
 export const fetchAddToCart = createAsyncThunk(
   'cart/add',
   async (id:string, {dispatch,rejectWithValue }) => {
     try {
       const request = await RequestCart.addToCart(id)
       if (request.status == 201 && request.data) {
-        dispatch(addCart(request.data))
+        dispatch(addCart(request.data.item))
+        dispatch(setTotalPrice(request.data.totalPrice))
       }
       
     } catch (error:any) {
@@ -34,7 +54,8 @@ export const fetchChangeAmount = createAsyncThunk(
       const request = await RequestCart.changeAmount(change)
       console.log(request,change)
       if (request.status == 201) {
-        dispatch(changeCart({id:change.cartId,changes:{amount:request.data as number}}))
+        dispatch(changeCart({ id: change.cartId, changes: { amount: request.data.amount as number } }))
+        dispatch(setTotalPrice(request.data.totalPrice))
       }
       
       
@@ -50,8 +71,9 @@ export const fetchRemoveCart = createAsyncThunk(
     try {
       
       const request = await RequestCart.removeCart(cartId)
-      if (request.status == 200) {
+      if (request.status == 200 && cartId === request.data.deletedId) {
         dispatch(removeCart(cartId))
+        dispatch(setTotalPrice(request.data.totalPrice))
       }
       
       
@@ -69,7 +91,22 @@ export const fetchDeleteCart = createAsyncThunk(
       const request = await RequestCart.deleteCart()
       if (request.status == 200) {
         dispatch(deleteCart())
+        dispatch(setTotalPrice(0))
       }
+      
+      
+    } catch (error:any) {
+      return rejectWithValue(error.response.data)
+    }
+  }
+)
+export const fetchOrderCart = createAsyncThunk(
+  'cart/order',
+  async (value:any, {dispatch,rejectWithValue }) => {
+    try {
+      
+      const request = await RequestCart.OrderCart(value)
+      console.log(request)
       
       
     } catch (error:any) {
@@ -82,16 +119,22 @@ const cartSlice = createSlice({
   name: 'cart',
   initialState:cartAdapter.getInitialState(CartEntities.getEntities),
   reducers: {
+    addAllCart:cartAdapter.addMany,
     addCart: cartAdapter.setOne,
     changeCart: cartAdapter.updateOne,
     removeCart: cartAdapter.removeOne,
     deleteCart: cartAdapter.removeAll,
     setAdress: (state, action) => {
       state.address = action.payload
+    },
+    setTotalPrice:(state, action) => {
+      state.totalPrice = action.payload
     }
   },
   
   
 })
-export const {addCart,changeCart,removeCart,deleteCart,setAdress} = cartSlice.actions
+export const {
+  addAllCart, addCart, changeCart, removeCart, deleteCart, setAdress,setTotalPrice
+} = cartSlice.actions
 export default cartSlice
